DOMEvents/tests/approved/EventListener.eventHandler.html

<!DOCTYPE html>
<meta charset=utf-8>
<title> EventLister member: handleEvent() </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<table id="table" border="1" style="display: none">
    <tbody id="table-body">
    <tr id="table-row">
        <td id="table-cell">Shady Grove</td>
        <td>Aeolian</td>
    </tr>
    <tr id="parent">
        <td id="target">Over the river, Charlie</td>
        <td>Dorian</td>
    </tr>
    </tbody>
</table>
<script>
test(function() {
    var t = this;
    var event_type = "foo";
    var target = document.getElementById("target");
    var handled = false;

    var event_listener = {
        "handleEvent": function(evt) {
            t.step(function() {
                handled = true;
                assert_equals(evt.type, event_type);
                assert_equals(evt.target, target);
                assert_equals(this, event_listener);
            }, this);
        }
    };

    var evt = document.createEvent("Event");
    evt.initEvent(event_type, true, true);
    
    target.addEventListener(event_type, event_listener, true);
    target.dispatchEvent(evt);
    target.removeEventListener(event_type, event_listener, true);

    assert_true(handled);
}, "handleEvent - This method shall be called whenever an event occurs of the event type for " +
                      "which the EventListener interface was registered.");
</script>
